Type hinting, cosmetics

+++ blocks/list_topic.php
@@ -62,7 +62,7 @@ function newbb_list_topic_show(array $options ): array
     $newbbConfig = newbbLoadConfig(); // load all newbb configs
 
     $topicRenderer            = new TopicRenderer();
-    $topicRenderer->userlevel = (int)$GLOBALS['xoopsUserIsAdmin'] ? 2 : is_object($GLOBALS['xoopsUser']); // Vistitor's level: 0 - anonymous; 1 - user; 2 - moderator or admin
+    $topicRenderer->userlevel = (int)$GLOBALS['xoopsUserIsAdmin'] ? 2 : (int)is_object($GLOBALS['xoopsUser']); // Vistitor's level: 0 - anonymous; 1 - user; 2 - moderator or admin
 
     $topicRenderer->force = true; // force against static vars for parse
